version 2.3.2: minor fixes in TrialBasic/TestBasic/AnswerDialogFragment/TestFragment: elapsed is now Long added some strings for new tests (TODO)
in TTC added:
 createFT_adaptive_VH_VARSPEED_VT_IPL
 createFT_adaptive_VH_VARSPEED_VPL_IT
 createFT_adaptive_VH_FIXSPEED

updated:
- chaquopy 14 -> 16
- compileSdkVersion 32 -> 34
- permissions
- gradle 7.4 -> 7.5

+++ psysuitecore/src/main/java/iit/uvip/psysuite/core/tests/tfi/AnswerDialogFragmentTFI.kt
@@ -164,7 +164,7 @@ class AnswerDialogFragmentTFI: DialogFragment()
         else    sendResult(curr_answer, elapsedms, TestBasic.EVENT_ANSWER_GIVEN)
     }
 
-    private fun sendResult(response: Int, elapsedTime: Int, response_id: Int) {
+    private fun sendResult(response: Int, elapsedTime:Long, response_id: Int) {
         if (targetFragment == null) return
 
         tts?.stop()
